LayoutTests/platform/win/plugins/window-region-is-set-to-clip-rect-expected.txt

CONSOLE MESSAGE: PLUGIN: PASS: Plugin's window's window region has been set as expected
CONSOLE MESSAGE: PLUGIN: PASS: Plugin's window's class does not have the CS_PARENTDC style

This tests that a plugin's window's window region is set to its clip rect.
